split_header: Reorganize headers
